requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai-whisper==20231117
python-multipart==0.0.12
pydantic==2.10.3
torch==2.3.1
numpy==1.26.4